chore(settings): remove unused UserSettings from meta/types
