+++ common/management/src/lib.rs
@@ -14,7 +14,6 @@
 
 mod cluster;
 mod role;
-mod setting;
 mod stage;
 mod udf;
 mod user;
@@ -23,8 +22,6 @@ pub use cluster::ClusterApi;
 pub use cluster::ClusterMgr;
 pub use role::RoleApi;
 pub use role::RoleMgr;
-pub use setting::SettingApi;
-pub use setting::SettingMgr;
 pub use stage::StageApi;
 pub use stage::StageMgr;
 pub use udf::UdfApi;
