+++ common/management/tests/it/main.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod cluster;
-mod setting;
 mod stage;
 mod udf;
 mod user;
