+++ common/meta/types/src/lib.rs
@@ -46,7 +46,6 @@ mod user_identity;
 mod user_info;
 mod user_privilege;
 mod user_quota;
-mod user_setting;
 mod user_stage;
 
 pub mod error_context;
@@ -191,5 +190,4 @@ pub use user_info::UserOptionFlag;
 pub use user_privilege::UserPrivilegeSet;
 pub use user_privilege::UserPrivilegeType;
 pub use user_quota::UserQuota;
-pub use user_setting::UserSetting;
 pub use user_stage::*;
