+++ query/src/sessions/session_settings.rs
@@ -22,15 +22,13 @@ use common_base::infallible::RwLock;
 use common_datavalues::prelude::*;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_meta_types::UserSetting;
 use itertools::Itertools;
 
-use crate::sessions::SessionContext;
-use crate::users::UserApiProvider;
 use crate::Config;
 
 #[derive(Clone)]
 enum ScopeLevel {
+    #[allow(dead_code)]
     Global,
     Session,
 }
@@ -53,7 +51,6 @@ pub struct SettingValue {
     // Default value of this setting.
     default_value: DataValue,
     user_setting: UserSetting,
-    // The scope of the setting is GLOBAL(metasrv) or SESSION.
     level: ScopeLevel,
     desc: &'static str,
 }
@@ -61,16 +58,10 @@ pub struct SettingValue {
 #[derive(Clone)]
 pub struct Settings {
     settings: Arc<RwLock<HashMap<String, SettingValue>>>,
-    user_api: Arc<UserApiProvider>,
-    session_ctx: Arc<SessionContext>,
 }
 
 impl Settings {
-    pub fn try_create(
-        conf: &Config,
-        session_ctx: Arc<SessionContext>,
-        user_api: Arc<UserApiProvider>,
-    ) -> Result<Settings> {
+    pub fn try_create(conf: &Config) -> Result<Settings> {
         let values = vec![
             // max_block_size
             SettingValue {
@@ -167,11 +158,7 @@ impl Settings {
             }
         }
 
-        let ret = Settings {
-            settings,
-            user_api,
-            session_ctx,
-        };
+        let ret = Settings { settings };
 
         // Overwrite settings from conf.
         {
@@ -287,42 +274,23 @@ impl Settings {
     }
 
     // Set u64 value to settings map, if is_global will write to metasrv.
-    fn try_set_u64(&self, key: &str, val: u64, is_global: bool) -> Result<()> {
+    fn try_set_u64(&self, key: &str, val: u64, _is_global: bool) -> Result<()> {
         let mut settings = self.settings.write();
         let mut setting = settings
             .get_mut(key)
             .ok_or_else(|| ErrorCode::UnknownVariable(format!("Unknown variable: {:?}", key)))?;
         setting.user_setting.value = DataValue::UInt64(val);
 
-        if is_global {
-            self.set_to_global(setting)?;
-        }
-
         Ok(())
     }
 
-    fn try_set_string(&self, key: &str, val: Vec<u8>, is_global: bool) -> Result<()> {
+    fn try_set_string(&self, key: &str, val: Vec<u8>, _is_global: bool) -> Result<()> {
         let mut settings = self.settings.write();
         let mut setting = settings
             .get_mut(key)
             .ok_or_else(|| ErrorCode::UnknownVariable(format!("Unknown variable: {:?}", key)))?;
         setting.user_setting.value = DataValue::String(val);
 
-        if is_global {
-            self.set_to_global(setting)?;
-        }
-
-        Ok(())
-    }
-
-    fn set_to_global(&self, setting: &mut SettingValue) -> Result<()> {
-        let tenant = self.session_ctx.get_current_tenant();
-        let _ = futures::executor::block_on(
-            self.user_api
-                .get_setting_api_client(&tenant)?
-                .set_setting(setting.user_setting.clone()),
-        )?;
-        setting.level = ScopeLevel::Global;
         Ok(())
     }
 
@@ -379,5 +347,23 @@ impl Settings {
         }
 
         Ok(())
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct UserSetting {
+    // The name of the setting.
+    pub name: String,
+
+    // The value of the setting.
+    pub value: DataValue,
+}
+
+impl UserSetting {
+    pub fn create(name: &str, value: DataValue) -> UserSetting {
+        UserSetting {
+            name: name.to_string(),
+            value,
+        }
     }
 }
